requirements.txt
APScheduler
fastapi
google-genai
seaborn
Pillow
SQLAlchemy
matplotlib
numpy
pandas
python-dateutil
python-dotenv
python-telegram-bot
pytz
requests
streamlit
uvicorn
psycopg2-binary
supabase